Creating charting script for TradingView

+++ src/main/java/com/zygne/stockalyze/domain/interactor/implementation/stats/StatisticsInteractor.java
@@ -18,7 +18,7 @@ public class StatisticsInteractor implements Interactor {
     @Override
     public void execute() {
 
-        int totalShares = 0;
+        long totalShares = 0;
         int sampleSize = data.size();
 
         for(Histogram e : data){
