+++ src/main/java/com/zygne/stockalyze/domain/model/VolumePriceLevel.java
@@ -3,9 +3,9 @@ package com.zygne.stockalyze.domain.model;
 public class VolumePriceLevel {
 
     public int price;
-    public int size;
+    public long size;
 
-    public VolumePriceLevel(int price, int size) {
+    public VolumePriceLevel(int price, long size) {
         this.size = size;
         this.price = price;
     }
